requirements-dev.txt
# Requirements for development and testing

# Base requirements
-r requirements.txt

# Testing
pytest~=8.3.4
pytest-cov~=6.0.0
pytest-mock~=3.14.0
pytest-timeout~=2.3.1
pytest-xdist~=3.6.1

# Type stubs
pandas-stubs~=2.2.3
types-requests~=2.32.0

# Development tools
pre-commit~=4.0.1
